fix: map autoProcessPaths values to platform path
On Windows, setting the autoProcessPaths react-native-unstyles/plugin config does not work, meaning importing 3rd party component libraries that use react-native-unistyles are not transformed. This is a fix for that, using the already built-in `toPlatformPath` function to ensure windows paths are converted correctly (i.e use backslashes instead of forward slashes).